backend/requirements.txt
# core dependencies
flask>=3.0.0
pyserial>=3.5
numpy>=1.26.0
scipy>=1.11.0
requests>=2.31.0

# environment variables (optional)
python-dotenv>=1.0.0

# development/testing (optional)
python-dotenv>=1.0.0